# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory Management System API
  description: |
    REST API for the Inventory Management System (IMS).

    Features include product catalog, multi-warehouse inventory, suppliers,
    purchase orders, sales orders, stock movements, reports, and audit logs.

    Real-time WebSocket endpoints (not part of this REST spec):
    - `ws://localhost:3000/ws/inventory`
    - `ws://localhost:3000/ws/products`
    - `ws://localhost:3000/ws/warehouses`
    - `ws://localhost:3000/ws/suppliers`
  version: 1.1.0
  contact:
    name: IMS API Support
servers:
  - url: /
    description: Current host (use for Swagger Try it out)
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: List products
      description: Returns all products, optionally filtered by category.
      operationId: listProducts
      parameters:
        - $ref: "#/components/parameters/CategoryQuery"
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Products]
      summary: Create product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
            example:
              name: Widget Pro
              sku: WGT-PRO-001
              description: Premium widget
              category: Electronics
              price: 299.99
              cost: 150.00
              unit: piece
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /products/search:
    get:
      tags: [Products]
      summary: Search products
      description: Search products by name, SKU, or description.
      operationId: searchProducts
      parameters:
        - name: q
          in: query
          required: true
          description: Search keyword
          schema:
            type: string
          example: widget
      responses:
        "200":
          description: Matching products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /products/{id}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      tags: [Products]
      summary: Get product by ID
      operationId: getProduct
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Products]
      summary: Update product
      operationId: updateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Products]
      summary: Delete product
      operationId: deleteProduct
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /warehouses:
    get:
      tags: [Warehouses]
      summary: List warehouses
      operationId: listWarehouses
      responses:
        "200":
          description: List of warehouses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Warehouses]
      summary: Create warehouse
      operationId: createWarehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarehouseRequest"
            example:
              name: North Warehouse
              location: Seattle, WA
              capacity: 15000
      responses:
        "201":
          description: Warehouse created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /warehouses/{id}:
    parameters:
      - $ref: "#/components/parameters/WarehouseId"
    get:
      tags: [Warehouses]
      summary: Get warehouse by ID
      operationId: getWarehouse
      responses:
        "200":
          description: Warehouse details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Warehouses]
      summary: Update warehouse
      operationId: updateWarehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWarehouseRequest"
      responses:
        "200":
          description: Updated warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Warehouses]
      summary: Delete warehouse
      operationId: deleteWarehouse
      responses:
        "200":
          description: Warehouse deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /inventory:
    get:
      tags: [Inventory]
      summary: Get inventory levels
      description: Returns current stock levels across warehouses.
      operationId: getInventory
      parameters:
        - name: warehouse_id
          in: query
          description: Filter by warehouse ID
          schema:
            type: integer
            format: uint
      responses:
        "200":
          description: Inventory records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /inventory/{productId}:
    parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: uint
    get:
      tags: [Inventory]
      summary: Get inventory for a product
      description: Returns stock levels for a product across all warehouses.
      operationId: getProductInventory
      responses:
        "200":
          description: Product inventory records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /inventory/adjust:
    post:
      tags: [Inventory]
      summary: Adjust stock
      description: |
        Manually adjust stock for a product in a warehouse.
        Creates a new inventory record if none exists; otherwise adds to the current quantity.
        Logs an ADJUST stock movement and audit entry.
      operationId: adjustInventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustInventoryRequest"
            example:
              product_id: 1
              warehouse_id: 1
              quantity: 100
              reason: Initial stock
      responses:
        "200":
          description: Inventory adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /inventory/low-stock:
    get:
      tags: [Inventory]
      summary: Get low-stock items
      description: Returns items where quantity is at or below minimum stock level.
      operationId: getLowStock
      responses:
        "200":
          description: Low-stock inventory records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LowStockResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /inventory/movements:
    get:
      tags: [Inventory]
      summary: Get stock movements
      description: Returns the latest 100 stock movement records.
      operationId: getStockMovements
      parameters:
        - name: product_id
          in: query
          description: Filter by product ID
          schema:
            type: integer
            format: uint
        - name: type
          in: query
          description: Movement type filter
          schema:
            type: string
            enum: [IN, OUT, ADJUST]
      responses:
        "200":
          description: Stock movement history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovementListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /suppliers:
    get:
      tags: [Suppliers]
      summary: List suppliers
      operationId: listSuppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Suppliers]
      summary: Create supplier
      operationId: createSupplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
            example:
              name: TechParts Inc
              contact_name: John Doe
              email: john@techparts.com
              phone: "+1-555-0123"
              address: 123 Supply St, NY
              rating: 4.5
      responses:
        "201":
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /suppliers/{id}:
    parameters:
      - $ref: "#/components/parameters/SupplierId"
    put:
      tags: [Suppliers]
      summary: Update supplier
      operationId: updateSupplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierRequest"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /purchase-orders:
    get:
      tags: [Purchase Orders]
      summary: List purchase orders
      operationId: listPurchaseOrders
      parameters:
        - $ref: "#/components/parameters/StatusQuery"
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Purchase Orders]
      summary: Create purchase order
      operationId: createPurchaseOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseOrderRequest"
            example:
              supplier_id: 1
              items:
                - product_id: 1
                  quantity: 500
                  unit_price: 15.00
      responses:
        "201":
          description: Purchase order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /purchase-orders/{id}/receive:
    parameters:
      - $ref: "#/components/parameters/PurchaseOrderId"
    put:
      tags: [Purchase Orders]
      summary: Receive purchase order
      description: Marks a purchase order as received and increases warehouse inventory.
      operationId: receivePurchaseOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceivePurchaseOrderRequest"
            example:
              warehouse_id: 1
      responses:
        "200":
          description: Purchase order received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceivePurchaseOrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders:
    get:
      tags: [Sales Orders]
      summary: List sales orders
      operationId: listOrders
      parameters:
        - $ref: "#/components/parameters/StatusQuery"
      responses:
        "200":
          description: List of sales orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Sales Orders]
      summary: Create sales order
      description: Creates an order and reduces inventory. Fails if stock is insufficient.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
            example:
              customer_name: ABC Corp
              customer_email: orders@abc.com
              items:
                - product_id: 1
                  warehouse_id: 1
                  quantity: 10
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /orders/{id}/status:
    parameters:
      - $ref: "#/components/parameters/OrderId"
    put:
      tags: [Sales Orders]
      summary: Update order status
      operationId: updateOrderStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
            example:
              status: shipped
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /reports/stock-summary:
    get:
      tags: [Reports]
      summary: Get stock summary report
      description: Returns stock value breakdown by warehouse and product.
      operationId: getStockSummary
      responses:
        "200":
          description: Stock summary report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockSummaryResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /orders/{id}:
    parameters:
      - $ref: "#/components/parameters/OrderId"
    get:
      tags: [Sales Orders]
      summary: Get order by ID
      operationId: getOrder
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      operationId: listCustomers
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: company
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      tags: [Customers]
      summary: Create customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
  /customers/{id}:
    parameters:
      - $ref: "#/components/parameters/CustomerId"
    get:
      tags: [Customers]
      summary: Get customer by ID
      operationId: getCustomer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
    put:
      tags: [Customers]
      summary: Update customer
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
    delete:
      tags: [Customers]
      summary: Delete customer
      operationId: deleteCustomer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /categories:
    get:
      tags: [Categories]
      summary: List categories
      operationId: listCategories
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListResponse"
    post:
      tags: [Categories]
      summary: Create category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
  /categories/{id}:
    parameters:
      - $ref: "#/components/parameters/CategoryId"
    get:
      tags: [Categories]
      summary: Get category
      operationId: getCategory
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
    put:
      tags: [Categories]
      summary: Update category
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
    delete:
      tags: [Categories]
      summary: Delete category
      operationId: deleteCategory
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /employees:
    get:
      tags: [Employees]
      summary: List employees
      operationId: listEmployees
      parameters:
        - name: role
          in: query
          schema: { type: string }
        - name: warehouse_id
          in: query
          schema: { type: integer }
        - name: active
          in: query
          schema: { type: string, enum: [true] }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
    post:
      tags: [Employees]
      summary: Create employee
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
  /employees/{id}:
    parameters:
      - $ref: "#/components/parameters/EmployeeId"
    get:
      tags: [Employees]
      summary: Get employee
      operationId: getEmployee
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
    put:
      tags: [Employees]
      summary: Update employee
      operationId: updateEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
    delete:
      tags: [Employees]
      summary: Delete employee
      operationId: deleteEmployee
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /transfers/{id}:
    parameters:
      - $ref: "#/components/parameters/TransferId"
    get:
      tags: [Transfers]
      summary: Get transfer by ID
      operationId: getTransfer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
...
  /reports/top-products:
    get:
      tags: [Reports]
      summary: Top selling products
      operationId: getTopProducts
      parameters:
        - name: limit
          in: query
          schema: { type: string, default: "10" }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopProductsResponse"
...
components:
  parameters:
...
    CategoryQuery:
      name: category
      in: query
      description: Filter products by category
      schema:
        type: string
...
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        text/plain:
          schema:
            type: string
            example: Failed to fetch products
  schemas:
...
    CreateProductRequest:
      type: object
      required: [name, sku, price]
      properties:
        name:
          type: string
        sku:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
          format: double
        cost:
          type: number
          format: double
        unit:
          type: string
...
    ProductListResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
...